+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -27,6 +27,7 @@
 #include "Rule18p7Check.h"
 #include "Rule18p8Check.h"
 #include "Rule19p2Check.h"
+#include "Rule8p10Check.h"
 
 namespace clang::tidy {
 namespace misrac {
@@ -72,6 +73,8 @@ public:
         "misrac-rule-18p8");
     CheckFactories.registerCheck<Rule19p2Check>(
         "misrac-rule-19p2");
+    CheckFactories.registerCheck<Rule8p10Check>(
+        "misrac-rule-8p10");
   }
 };